pyproject.toml
[tool.poetry]
name = "pyuhoo"
version = "0.0.3"
description = "Python API for talking to uHoo consumer API"
authors = ["Christopher Sacca <csacca@csacca.net>"]
repository = "https://github.com/csacca/pyuhoo"
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
pyuhoo-cli = "pyuhoo.cli:cli"

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.7.4"
click = "^8.0.1"
pycryptodome = "^3.10.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
black = "^21.7b0"
isort = "^5.9.3"
mypy = "^0.910"
pre-commit = "^2.14.0"
safety = "^1.10.3"
pytest = "^6.2.4"
pytest-aiohttp = "^0.3.0"

[tool.isort]
profile = "black"

[tool.mypy]
pretty = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
